+++ CHANGELOG.md
@@ -21,6 +21,7 @@ var ecdsa = require('ecdsa')('secp256r1')
 ```
 
 * added `deterministicGenerateK()`, RFC 6979. See: https://github.com/cryptocoinjs/ecdsa/issues/4.
+* added `recoverPubKey()` from BitcoinJS
 
 
 0.4.1 / 2014-04-14